requirements.txt
cassandra-driver

